refactor<rve>: convert to MynaApp implementation
